refactor: ResultChartItem 컴포넌트로 차트 항목 분리
