+++ src/components/form/result/ResultDetail.tsx
@@ -3,29 +3,12 @@ import { css } from '@emotion/react'
 import { Text } from '@/components/shared/text'
 import { SingleAlbumData } from '@/models/album'
 import { useGetTrackCounts } from '@/hooks/useGetTrackCounts'
-import { colors } from '@/styles/colors'
+import ResultChartItem from './ResultChartItem'
 
 interface ResultDetailProps {
   album: SingleAlbumData
   voteCount: number
 }
-
-const getRandomColor = (() => {
-  const colorOptions: (keyof typeof colors)[] = [
-    'pink',
-    'yellow',
-    'blue',
-    'green',
-    'purple',
-  ]
-  let startingColorIndex = Math.floor(Math.random() * colorOptions.length)
-
-  return (index: number) => {
-    const colorIndex = (index + startingColorIndex) % colorOptions.length
-    const selectedColor = colorOptions[colorIndex]
-    return colors[selectedColor]
-  }
-})()
 
 export default function ResultDetail({ album, voteCount }: ResultDetailProps) {
   const { formId } = useParams()
@@ -39,37 +22,15 @@ export default function ResultDetail({ album, voteCount }: ResultDetailProps) {
         <Text>최애곡은?</Text>
       </div>
       <ul>
-        {tracks.map((track, index) => {
-          const trackVoteCount = trackCounts[track.track_number] || 0
-          let trackPercentage
-          if (trackVoteCount === 0) {
-            trackPercentage = 0
-          } else {
-            trackPercentage = (trackVoteCount / voteCount) * 100
-          }
-          if (Number.isInteger(trackPercentage)) {
-            trackPercentage = Math.round(trackPercentage)
-          } else {
-            trackPercentage = parseFloat(trackPercentage.toFixed(1))
-          }
-
-          const color = getRandomColor(index)
-
-          return (
-            <li key={track.id}>
-              <div css={listInfoStyles}>
-                <Text css={trackNameStyles}>{track.name}</Text>
-                <div>
-                  <Text css={voteCountStyles} variant={'detailStrong'}>
-                    {trackVoteCount}
-                  </Text>
-                  <Text variant={'detail'}>{trackPercentage}%</Text>
-                </div>
-              </div>
-              <div css={chartBarStyles(trackPercentage, color)}></div>
-            </li>
-          )
-        })}
+        {tracks.map((track, index) => (
+          <ResultChartItem
+            key={track.id}
+            track={track}
+            trackVoteCount={trackCounts[track.track_number] || 0}
+            voteCount={voteCount}
+            index={index}
+          />
+        ))}
       </ul>
     </section>
   )
@@ -77,38 +38,4 @@ export default function ResultDetail({ album, voteCount }: ResultDetailProps) {
 
 const titleStyles = css`
   padding: 6px 0;
-`
-
-const listInfoStyles = css`
-  display: flex;
-  justify-content: space-between;
-  align-items: end;
-  gap: 10px;
-  padding: 6px 0;
-`
-
-const trackNameStyles = css`
-  overflow: hidden;
-  text-overflow: ellipsis;
-`
-
-const voteCountStyles = css`
-  margin-right: 4px;
-`
-
-const chartBarStyles = (trackPercentage: number, color: string) => css`
-  position: relative;
-  width: 100%;
-  height: 24px;
-  background-color: ${colors.gray200};
-  border-radius: 2px;
-
-  ::after {
-    content: '';
-    position: absolute;
-    inset: 0;
-    width: ${trackPercentage}%;
-    background-color: ${color};
-    border-radius: 2px;
-  }
 `
